build+tests: make the tree compile and pass on macOS (Apple clang)
The README claims 'Platform: Linux, macOS (partial)' but the tree did
not compile on Darwin at all:

- -D_POSIX_C_SOURCE hides O_NOFOLLOW/strcasestr on Darwin -> every
  symlink-protected open() failed to compile. New PLATFORM_CFLAGS adds
  -D_DARWIN_C_SOURCE on macOS, applied on every compile line including
  the literal-flag fuzz/coverage/ASan recipes.
- check-warnings used GCC-only spellings (-Wformat-*=2,
  -Wstringop-overread) that clang rejects with an unknown-option
  warning per file; the strict set is now compiler-conditional.
- -pie is a no-op on macOS (always PIE) and clang warned on every link;
  PIE_LDFLAGS is empty on Darwin.
- make static cannot work on macOS (no static libc); it now fails with
  an explicit message.
- hlse_audit.c: dropped dead nameserver_count; hlse_core.c: explicit
  AuditSeverity->int cast (clang -Wsign-conversion).

Test suite portability:
- p111 invoked the binary via a hardcoded /home/user/HLSE/hlse_core --
  it could only pass on one machine. The suite now captures HLSE_ROOT
  and uses it inside the temp-dir subshells.
- p53/p61 assumed the host sudoers has a NOPASSWD entry; on a hardened
  host nothing reaches HIGH and they failed spuriously. They now probe
  --json audit once and SKIP when absent (existing SKIP convention).

Verified on macOS (Apple clang 21): make, check-warnings (0 warnings,
CLI + lib), make test (all suites green, CLI 776/0, 3 audit checks
SKIPped), asan-test clean, fuzz 6x100K 0 crashes, server-check 14/14.

Devin-Session: https://app.devin.ai/sessions/a662875256294db2a70742ed489e0208

+++ hlse_core.c
@@ -9155,7 +9155,7 @@ main(int argc, char **argv) {
                 const char *sev_str[] = {
                     "PASS", "INFO", "LOW", "MED", "HIGH", "CRIT"
                 };
-                int s = av.findings[i].severity;
+                int s = (int)av.findings[i].severity;
                 if (s < 0 || s > 5) s = 0;
                 printf("  [%4s] %s\n", sev_str[s],
                        av.findings[i].description);
